requirements.txt
# ─────────────────────────────────────────────────────────────────
# SmartAgri Dashboard — Python 3.11 requirements
# Install:  pip install -r requirements.txt
# Run:      streamlit run app.py
# ─────────────────────────────────────────────────────────────────

# ── Web framework ─────────────────────────────────────────────────
streamlit>=1.28.0,<1.32.0

# ── Mapping ───────────────────────────────────────────────────────
folium>=0.14.0,<0.16.0
streamlit-folium>=0.15.0,<0.21.0

# ── GIS  (all have prebuilt wheels for Python 3.11 Windows) ──────
geopandas>=0.13.0,<0.14.0
shapely>=2.0.0,<2.1.0
pyproj>=3.5.0,<3.7.0
fiona>=1.9.0,<1.10.0
mapclassify>=2.5.0,<2.7.0

# ── Google Earth Engine ───────────────────────────────────────────
earthengine-api>=0.1.380,<0.1.400

# ── Data / ML ─────────────────────────────────────────────────────
numpy>=1.24.0,<1.27.0
pandas>=2.0.0,<2.2.0
scikit-learn>=1.2.0,<1.4.0
joblib>=1.2.0,<1.4.0

# ── Visualisation ─────────────────────────────────────────────────
plotly>=5.14.0,<5.20.0
matplotlib>=3.7.0,<3.9.0

# ── Image / raster ────────────────────────────────────────────────
Pillow>=10.0.0,<10.3.0
rioxarray>=0.14.0,<0.16.0
xarray>=2023.1.0,<2024.3.0

# ── Utilities ─────────────────────────────────────────────────────
pyyaml>=6.0,<6.1
requests>=2.29.0,<2.32.0
tqdm>=4.65.0,<4.67.0
